Computed dmg by player, teammates, enemies. Sth's wrong.

+++ WebApplication1/Models/DamageDealt.cs
@@ -6,10 +6,10 @@ namespace WebApplication1.Models
     public class DamageDealt
     {
         [DataMember]
-        public int averageDmgToChampions;
+        public double averageDmgToChampions;
 
         [DataMember]
-        public int averageDmgToTurrets;
+        public double averageDmgToTurrets;
 
         public DamageDealt() { }
     }
